完成管理员评价获取删除，校验bug修改

+++ src/main/java/com/wsl/shoppingKill/service/UserService.java
@@ -35,7 +35,7 @@ public interface UserService extends IService<User> {
      * @param id:
      * @return boolean
      **/
-    boolean delUserInfo(Integer id);
+    boolean delUserInfo(Long id);
 
 
 }
